separated js from ts
